Interface

Alternates - iterator for quantum numbers

ExtendedKronigPennyMatrix.Alternates — Type

Alternates(nmax)

is an iterator to generate the series $\{0, 1, -1, 2, -2, \ldots \}$ up to nmax as an ordering of quantum numbers.

collect(Alternates(0)) => [0]
collect(Alternates(1)) => [0, 1, -1]
collect(Alternates(2)) => [0, 1, -1, 2, -2]

Base.iterate — Function

Base.iterate(alt::Alternates, state::Int = 1)

Potential

ExtendedKronigPennyMatrix.Potential — Type

Potential

is an abstraction of potential including

potential height, and/or

other parameters depending on specific potential.

A subtype of Potential is expected to possess following methods:

get_potential(<:Potential)

returns a function to evaluate potential value as a position.

Model

ExtendedKronigPennyMatrix.Model — Type

Model

is an abstraction of model including

potential <: Potential

nmax: maximum of quantum number

mmax: size of Hailtonian matrix

Ka : wavenumber multiplied by a, period

qnum <: Alternates

and other parameters depending on specific potential

A subtype of model is expected to possess following methods:

update!(<:Model, ...)

updates parameters in model.

Kronig-Penny model

ExtendedKronigPennyMatrix.KronigPennyPotential — Type

mutable struct KronigPennyPotential <: Potential

KronigPennyPotential(v0, ρ) # Default constructor

holds parameters of Kronig-Penny potential.

Fields

v0 : potential height in units of $E_{1}^{(0)}$

ρ : barrier width in units of period $a$, $\rho = \dfrac{b}{a}$, $0 < \rho < 1$

Note that a position $x$ is expressed in units of $a$ throughout this package.

The constructor KronigPennyPotential(v0, ρ) confirms that 0 ≤ ρ ≤ 1, otherwise throws an error.

ExtendedKronigPennyMatrix.get_potential — Method

get_potential(::KronigPennyPotential)

returns a function $v$ to evaluate potential $v(x)$ as a position $x$, such that:

\[\begin{aligned} v(x) & = \begin{cases} v_{0} & \text{inside well, i.e.,} \dfrac{1-\rho}{2} \le \dfrac{x}{a} \le \dfrac{1+\rho}{2}, \\ 0 & \text{outside well}\end{cases} \\ v(x+a) &= v(x) \end{aligned}\]

Note that a position $x$ is expressed in units of $a$ throughout this package.

ExtendedKronigPennyMatrix.KronigPennyModel — Type

mutable struct KronigPennyModel <: Model

holds parameters of Kronig-Penny model.

Fields given to the constructor:

potential

nmax

Ka : wavenumber multiplied by a, period

Fields calculated by constructor:

mmax : size of Hamiltonian matrix

qnum : iterator of quantum numbers

Fields calculated by constructor and update! methods:

hnm : hamiltonian matrix

ExtendedKronigPennyMatrix.KronigPennyModel — Method

KronigPennyModel(::KronigPennyPotential, nmax, Ka)

is a constructor of Kronig-Penny model, and defines other fields: qnum, nmax, and hnm

Mandantory parameters:

pot : potential

nmax : maximum of quantum numbers

Ka : wavenumber multiplied by a, period

ExtendedKronigPennyMatrix.update! — Method

update!(model_lem::KronigPennyModel; Ka=nothing)

updates Kronnig-Penny model.

Optional parameters

Ka : wavenumber multiplied by a, period

ExtendedKronigPennyMatrix.constuctMatrix — Method

constuctMatrix(model_lem::KronigPennyModel)

computes and fills Hamiltonian matrix fields hnm in model_lem.

\[h_{nm} = \begin{cases} \left(2n + \dfrac{Ka}{\pi}\right)^{2} + v_{0} (1-\rho) & \text{for}\; n = m\;\text{(diagonal elements)} \\ v_{0} \dfrac{(-1)^{m-n+1}}{\pi} \dfrac{\sin \pi(m-n)\rho}{m-n} & \text{for}\; n \neq m\;\text{(off-diagonal elements)}\end{cases}\]

Alphabetical Index

ExtendedKronigPennyMatrix.Alternates

ExtendedKronigPennyMatrix.KronigPennyModel

ExtendedKronigPennyMatrix.KronigPennyModel

ExtendedKronigPennyMatrix.KronigPennyPotential

ExtendedKronigPennyMatrix.Model

ExtendedKronigPennyMatrix.Potential

Base.iterate

ExtendedKronigPennyMatrix.constuctMatrix

ExtendedKronigPennyMatrix.get_potential

ExtendedKronigPennyMatrix.update!